+++ pages/COMPLETED/agent/[id].js
@@ -4,11 +4,11 @@
  * @returns The return value of the function is an object with a props property.
  */
 import React from "react";
-import NavbarThree from "../../layout/headers/NavbarOne";
+import NavbarThree from "../../../layout/headers/NavbarOne";
 
-import FooterThree from "../../layout/footers/FooterThree";
-import Breadcrumb from "../../layout/Breadcrumb/Breadcrumb";
-import BodyContent from "../../components/agent/agentProfile";
+import FooterThree from "../../../layout/footers/FooterThree";
+import Breadcrumb from "../../../layout/Breadcrumb/Breadcrumb";
+import BodyContent from "../../../components/agent/agentProfile";
 
 const AgentProfile = () => {
   return (
