+++ pages/COMPLETED/home/index.js
@@ -4,10 +4,9 @@
  */
 
 import { useEffect } from "react";
-import BodyContent from "../../components/home/slider-filter-search";
-import { ConfigDB } from "../../config/themeCustomizerConfig";
-import FooterOne from "../../layout/footers/FooterThree";
-import NavbarOne from "../../layout/headers/NavbarOne";
+import BodyContent from "../../../components/home/slider-filter-search";
+import FooterOne from "../../../layout/footers/FooterThree";
+import NavbarOne from "../../../layout/headers/NavbarOne";
