--- pages/agent/submit-property.js
@@ -1,21 +1,0 @@
-import React from "react";
-
-import BodyContent from "../../components/agent/submitProperty";
-import Breadcrumb from "../../layout/Breadcrumb/Breadcrumb";
-import FooterThree from "../../layout/footers/FooterThree";
-import NavbarThree from "../../layout/headers/NavbarOne";
-
-
-
-const SubmitProperty = () => {
-  return (
-    <>
-      <NavbarThree />
-      <Breadcrumb />
-      <BodyContent />
-      <FooterThree />
-    </>
-  );
-};
-
-export default SubmitProperty;
