--- pages/pages/other-pages/about-us-2.js
@@ -1,21 +1,0 @@
-import React from "react";
-
-import NavbarThree from "../../../layout/headers/NavbarOne";
-
-
-import FooterThree from "../../../layout/footers/FooterThree";
-import BodyContent from "../../../components/pages/otherPages/aboutUs2";
-import BreadCrumb2 from "../../../layout/Breadcrumb/BreadCrumb2";
-
-const AboutUs2 = () => {
-  return (
-    <>
-      <NavbarThree />
-      <BreadCrumb2 />
-      <BodyContent />
-      <FooterThree />
-    </>
-  );
-};
-
-export default AboutUs2;
